+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/apis/queryBuilderFetcher.ts
@@ -5,6 +5,8 @@ import {
   findValues,
   FindKeysOptions,
   FindValuesOptions,
+  TruncatedResult,
+  FindBucketsOptions,
 } from './fluxQueries'
 
 // Types
@@ -12,7 +14,7 @@ import {CancelBox} from 'src/types/promises'
 import {Source} from 'src/types'
 import {BuilderTagsType} from '../types'
 
-type CancelableQuery = CancelBox<string[]>
+type CancelableQuery = CancelBox<string[] | TruncatedResult<string[]>>
 
 function tagSelectionKey(tags: BuilderTagsType[]): any[] {
   return tags.map(x => ({
@@ -25,25 +27,42 @@ class QueryBuilderFetcher {
   private findBucketsQuery?: CancelableQuery
   private findKeysQueries: Array<CancelableQuery | undefined> = []
   private findValuesQueries: Array<CancelableQuery | undefined> = []
-  private findKeysCache: {[key: string]: string[]} = {}
-  private findValuesCache: {[key: string]: string[]} = {}
-  private findBucketsCache: {[key: string]: string[]} = {}
+  private findKeysCache: {
+    [key: string]: TruncatedResult<string[]> & {limit: number}
+  } = {}
+  private findValuesCache: {
+    [key: string]: TruncatedResult<string[]> & {limit: number}
+  } = {}
+  private findBucketsCache: {
+    [key: string]: TruncatedResult<string[]> & {limit: number}
+  } = {}
 
-  public async findBuckets(source: Source): Promise<string[]> {
+  public async findBuckets(
+    source: Source,
+    options: FindBucketsOptions = {limit: Number.MAX_SAFE_INTEGER}
+  ): Promise<TruncatedResult<string[]>> {
     this.cancelFindBuckets()
 
-    const cachedResult = this.findBucketsCache[source.id]
+    const cacheKey = source.id
+    const cachedResult = this.findBucketsCache[cacheKey]
 
-    if (cachedResult) {
-      return Promise.resolve(cachedResult)
+    if (cachedResult && options.limit <= cachedResult.limit) {
+      return Promise.resolve({
+        result: cachedResult.result,
+        truncated: cachedResult.truncated,
+      })
     }
 
-    const pendingResult = findBuckets(source)
+    const pendingResult = findBuckets(source, options)
     this.findBucketsQuery = pendingResult
 
     pendingResult.promise
-      .then(result => {
-        this.findBucketsCache[source.id] = result
+      .then(t => {
+        this.findBucketsCache[cacheKey] = {
+          result: t.result,
+          truncated: t.truncated,
+          limit: options.limit,
+        }
       })
       .catch(() => {})
 
@@ -60,10 +79,10 @@ class QueryBuilderFetcher {
   public async findKeys(
     tagIndex: number,
     options: FindKeysOptions
-  ): Promise<string[]> {
+  ): Promise<TruncatedResult<string[]>> {
     this.cancelFindKeys(tagIndex)
 
-    const {source, tagsSelections, ...rest} = options
+    const {source, tagsSelections, limit, ...rest} = options
     const cacheKey = JSON.stringify({
       id: source.id,
       tags: tagSelectionKey(tagsSelections),
@@ -71,17 +90,24 @@ class QueryBuilderFetcher {
     })
     const cachedResult = this.findKeysCache[cacheKey]
 
-    if (cachedResult) {
-      return Promise.resolve([...cachedResult])
+    if (cachedResult && limit <= cachedResult.limit) {
+      return Promise.resolve({
+        result: [...cachedResult.result],
+        truncated: cachedResult.truncated,
+      })
     }
 
     const pendingResult = findKeys(options)
 
     this.findKeysQueries[tagIndex] = pendingResult
 
-    pendingResult.promise = pendingResult.promise.then(result => {
-      this.findKeysCache[cacheKey] = [...result]
-      return result
+    pendingResult.promise = pendingResult.promise.then(t => {
+      this.findKeysCache[cacheKey] = {
+        result: [...t.result],
+        truncated: t.truncated,
+        limit,
+      }
+      return t
     })
 
     return pendingResult.promise
@@ -97,14 +123,14 @@ class QueryBuilderFetcher {
   public async findValues(
     tagIndex: number,
     options: FindValuesOptions
-  ): Promise<string[]> {
+  ): Promise<TruncatedResult<string[]>> {
     this.cancelFindValues(tagIndex)
     if (!options.key) {
       // return no values for no key
-      return []
+      return {result: [], truncated: false}
     }
 
-    const {source, tagsSelections, ...rest} = options
+    const {source, tagsSelections, limit, ...rest} = options
     const cacheKey = JSON.stringify({
       id: source.id,
       tags: tagSelectionKey(tagsSelections),
@@ -112,17 +138,24 @@ class QueryBuilderFetcher {
     })
     const cachedResult = this.findValuesCache[cacheKey]
 
-    if (cachedResult) {
-      return Promise.resolve([...cachedResult])
+    if (cachedResult && limit <= cachedResult.limit) {
+      return Promise.resolve({
+        result: [...cachedResult.result],
+        truncated: cachedResult.truncated,
+      })
     }
 
     const pendingResult = findValues(options)
 
     this.findValuesQueries[tagIndex] = pendingResult
 
-    pendingResult.promise = pendingResult.promise.then(result => {
-      this.findValuesCache[cacheKey] = [...result]
-      return result
+    pendingResult.promise = pendingResult.promise.then(t => {
+      this.findValuesCache[cacheKey] = {
+        result: [...t.result],
+        truncated: t.truncated,
+        limit,
+      }
+      return t
     })
 
     return pendingResult.promise