+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/reducers/tags.ts
@@ -1,6 +1,7 @@
 import {TagSelectorAction} from '../actions/tags'
 import {TagSelectorState} from '../types'
 import {BuilderAggregateFunctionType, RemoteDataState} from 'src/types'
+import {FQB_RESULTS_LIMIT} from '../apis/fluxQueries'
 
 export const initialState: TagSelectorState[] = []
 export function initialSelectorState(
@@ -13,11 +14,15 @@ export function initialSelectorState(
     keys: [],
     keysSearchTerm: '',
     keysStatus: RemoteDataState.NotStarted,
+    keysTruncated: false,
+    keysLimit: FQB_RESULTS_LIMIT,
     tagKey: '',
     values: [],
     valuesSearchTerm: '',
     valuesStatus: undefined,
+    valuesTruncated: false,
     tagValues: [],
+    valuesLimit: FQB_RESULTS_LIMIT,
   }
 }
 
@@ -75,6 +80,11 @@ const aggregationReducer = (
         valuesSearchTerm: action.payload.term,
       }))
     }
+    case 'FQB_TAG_INC_KEYS_LIMIT': {
+      return changeTagSelector(state, action.payload.tagIndex, tagState => ({
+        keysLimit: (tagState.keysLimit || 0) + FQB_RESULTS_LIMIT,
+      }))
+    }
     case 'FQB_TAG_SELECT_KEY': {
       return changeTagSelector(state, action.payload.tagIndex, () => ({
         tagKey: action.payload.key,
@@ -95,6 +105,11 @@ const aggregationReducer = (
     case 'FQB_TAG_SEARCH_VALUES': {
       return changeTagSelector(state, action.payload.tagIndex, () => ({
         valuesStatus: RemoteDataState.Loading,
+      }))
+    }
+    case 'FQB_TAG_INC_VALUES_LIMIT': {
+      return changeTagSelector(state, action.payload.tagIndex, tagState => ({
+        valuesLimit: (tagState.valuesLimit || 0) + FQB_RESULTS_LIMIT,
       }))
     }
     case 'FQB_TAG_KEY_STATUS': {
@@ -127,12 +142,14 @@ const aggregationReducer = (
       return changeTagSelector(state, action.payload.tagIndex, () => ({
         keys: action.payload.keys,
         keysStatus: RemoteDataState.Done,
+        keysTruncated: action.payload.truncated,
       }))
     }
     case 'FQB_TAG_VALUES': {
       return changeTagSelector(state, action.payload.tagIndex, () => ({
         values: action.payload.values,
         valuesStatus: RemoteDataState.Done,
+        valuesTruncated: action.payload.truncated,
       }))
     }
   }